bug fix: "correct" the sum-of-Fins iso
We needed an isomorphism (Fin m + Fin n <-> Fin (m + n)).  In some
sense any isomorphism will do, so I simply implemented the isomorphism
which was easiest to prove.  However, it turns out to be a "funny" one,
which in particular sends (Left i) to (n + i) and (Right i) to i.
It turns out, however, that this really matters, because when we
create Sp values we implicitly assume certain isomorphisms by the way
we combine element vectors.  So in particular, when we form a product
structure we simply concatenate the two element vectors, which
implicitly assumes an isomorphism which sends (Left i) to i and (Right
i) to (m + i).

This patch does some swapping and so on to construct the "correct"
isomorphism.  We also no longer export finNSum and finNSum' from the
module since they now correspond to a different isomorphism (the old
one which was easy to implement) and I don't want anyone to use them
by mistake.  They are really just an implementation detail anyway.

There may be a similar bug in the way we handle (Fin m * Fin n <->
Fin (m * n)) which might affect composition.  At the very least it
would affect our Array code.  I will take a look at that soon.